Extract API route handlers from apps to next package
The four session API route handlers (sessions, session by public id,
session-data, sessions:share) were byte-identical across CZ/SK/MK.
They move to @kalkulacka-one/next under api/routes/*, and the app
route files become one-line re-exports.

The session image route (OG/Instagram share images) differed only in
five strings, so it becomes a createSessionImageRoute() factory; each
app passes its election name, app title, and share texts in its route
file. This makes the per-election strings (TENANT-013) a visible
per-app config instead of edits buried in a 550-line copy.
