+++ apps/www.volebnikalkulacka.cz/app/api/calculators/[calculator-id]/sessions:share/route.ts
@@ -1,61 +1,1 @@
-import { HttpError, NotFoundError, UnauthorizedError } from "@kalkulacka-one/app";
-import { prisma } from "@kalkulacka-one/database";
-import { getEmbedNameFromRequest, getSessionCookie, getSessionFromRequest } from "@kalkulacka-one/next/session/server";
-
-import type { NextRequest } from "next/server";
-
-export async function POST(request: NextRequest, { params }: { params: Promise<{ "calculator-id": string }> }) {
-  try {
-    const { "calculator-id": calculatorId } = await params;
-
-    const embedName = getEmbedNameFromRequest(request);
-    const cookieData = await getSessionCookie({ embedName });
-    const sessionId = cookieData?.id || getSessionFromRequest(request);
-    if (!sessionId) {
-      return new UnauthorizedError("Session required").toResponse();
-    }
-
-    const session = await prisma.calculatorSession.findUnique({
-      where: {
-        sessionId_calculatorId: {
-          sessionId,
-          calculatorId,
-        },
-      },
-    });
-
-    if (!session) {
-      return new NotFoundError("Session not found for this calculator").toResponse();
-    }
-
-    if (session.publicId) {
-      return Response.json({
-        publicId: session.publicId,
-      });
-    }
-
-    const publicId = crypto.randomUUID();
-
-    await prisma.calculatorSession.update({
-      where: {
-        id: session.id,
-      },
-      data: {
-        publicId,
-      },
-    });
-
-    const ogImageUrl = new URL(`/api/images/sessions/${publicId}/opengraph`, request.url);
-    void fetch(ogImageUrl.toString()).catch(() => {});
-
-    return Response.json({
-      publicId,
-    });
-  } catch (error) {
-    if (error instanceof HttpError) {
-      return error.toResponse();
-    }
-
-    throw error;
-  }
-}
+export { POST } from "@kalkulacka-one/next/api/routes/sessions-share";
